Use size_t when referring to length and indices

+++ include/sddf/util/bitarray.h
@@ -23,7 +23,7 @@
 #define BITS_2_WORDS64(bits) (((bits)+63)/64)
 
 typedef struct bitarray {
-    uint64_t num_bits; /* Total number of bits */
+    size_t num_bits; /* Total number of bits */
     uint64_t *words; /* Word array */
 } bitarray_t;
 
@@ -34,7 +34,7 @@ typedef struct bitarray {
  * @param words pointer to the word array.
  * @param num_of_bits number of bits in the word array.
  */
-void bitarray_init(bitarray_t *bitarr, uint64_t *words, uint64_t num_bits);
+void bitarray_init(bitarray_t *bitarr, uint64_t *words, size_t num_bits);
 
 /**
  * Get the value of a specific bit in the bit array.
@@ -44,7 +44,7 @@ void bitarray_init(bitarray_t *bitarr, uint64_t *words, uint64_t num_bits);
  *
  * @return the value of the bit (0 or 1).
  */
-char bitarray_get_bit(bitarray_t *bitarr, uint64_t index);
+char bitarray_get_bit(bitarray_t *bitarr, size_t index);
 
 /**
  * Set all the bits to 0 or 1 in a specific region of the bit array.
@@ -54,7 +54,7 @@ char bitarray_get_bit(bitarray_t *bitarr, uint64_t index);
  * @param len length of the region.
  * @param value true sets the bits to 1, false sets the bits to 0.
  */
-void bitarray_set_region(bitarray_t *bitarr, uint64_t start, uint64_t len, bool value);
+void bitarray_set_region(bitarray_t *bitarr, size_t start, size_t len, bool value);
 
 /**
  * Toggle all the bits in a specific region of the bit array.
@@ -63,7 +63,7 @@ void bitarray_set_region(bitarray_t *bitarr, uint64_t start, uint64_t len, bool 
  * @param start starting index of the region.
  * @param len length of the region.
  */
-void bitarray_toggle_region(bitarray_t *bitarr, uint64_t start, uint64_t len);
+void bitarray_toggle_region(bitarray_t *bitarr, size_t start, size_t len);
 
 /**
  * Count the number of contiguous bits set to 0 or 1 starting from the start bit.
@@ -74,4 +74,4 @@ void bitarray_toggle_region(bitarray_t *bitarr, uint64_t start, uint64_t len);
  *
  * @return number of contiguous bits starting from start having the value provided.
  */
-uint64_t bitarray_count_bits(bitarray_t *bitarr, uint64_t start, bool is_set);
+size_t bitarray_count_bits(bitarray_t *bitarr, size_t start, bool is_set);
